requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0 
matplotlib>=3.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
scipy>=1.7.0 
face_alignment>=1.2.2